runware_integration/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0
